Changed smell level to be int

+++ include/icsv/detector/evaluation_center.hpp
@@ -15,7 +15,7 @@ class SmellEvaluator;
 
 class EvaluationCenter final {
 public:
-  using SmellLevel = double;
+  using SmellLevel = int;
 
 public:
   static auto Get() -> EvaluationCenter&;
